% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ess}
\alias{ess}
\title{ESS sample data}
\format{
A data frame with 143 rows (observations) and 38 columns (variables):
\describe{
\item{essround}{ESS round}
\item{idno}{Respondent's identification number}
\item{cntry}{Country}
\item{gndr}{Gender}
\item{agea}{Age of respondent (calculated)}
\item{edlvdno}{Highest level of education (Norwegian system)}
\item{mainact}{Main activity last 7 days}
\item{mbtru}{Member of trade union}
\item{hinctnta}{Household's total net income, all sources, in deciles}
\item{tvtot}{Total time spent watching TV}
\item{ppltrst}{Most people can be trusted vs. you can't be too careful}
\item{vote}{Voted in last national election}
\item{stflife}{How satisfied with life as a whole?}
\item{gincdif}{Government should reduce differences in incomes}
\item{freehms}{Gays & lesbians should be free to live as they wish}
\item{imwbcnt}{Immigrants make country worse or better place}
\item{happy}{How happy are you?}
\item{health}{Subjective general health}
\item{ctzcntr}{Citizen of country}
\item{brncntr}{Born in country?}
\item{height}{Height of respondent (cm)}
\item{weight}{Weight of respondent (kg)}
}
}
\source{
ESS Round 7: European Social Survey Round 7 Data (2014).
Data file edition 2.2. NSD - Norwegian Centre for Research Data, Norway –
Data Archive and distributor of ESS data for ESS ERIC.
\url{http://dx.doi.org/10.21338/NSD-ESS7-2014}; distributed under the
\href{https://creativecommons.org/licenses/by-nc-sa/4.0/}{CC BY-NC-SA 4.0 license}.
}
\usage{
ess
}
\description{
A subset of the 2014 (round 7) European Social Survey, limited to
143 respondents from Norway and 22 variables.
}
\keyword{datasets}
