% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krysstabell.R
\name{krysstabell}
\alias{krysstabell}
\title{Create a cross table showing column percentages}
\usage{
krysstabell(dataset, rowvar, colvar, export = NULL)
}
\arguments{
\item{dataset}{Your dataset (a data.frame-type object).}

\item{rowvar}{The row-variable.}

\item{colvar}{The column-variable.}

\item{export}{Should the result be made export-ready (TRUE/FALSE)?}
}
\value{
A matrix or, if export function is switched on (export=TRUE), a
printed out table formatted for easy export to MS Word.
}
\description{
This function allows you to cross-tabulate two variables in your dataset.
The table will show column percentages. The result can be printed in an
"export-ready" fashion, meaning it can be copied and pasted into a Word
document and there transformed into a publication-quality table. The
functions will provide warnings in some cases where user selections
might be problematic.
}
\examples{
\dontrun{
# Load mtcars dataset
data(mtcars)

# Simple example
krysstabell(dataset = mtcars,
rowvar = "carb",
colvar = "gear")

# For export
krysstabell(dataset = mtcars,
rowvar = "carb",
colvar = "gear",
export = TRUE)
}

}
