% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oppsum_grupp.R
\name{oppsum_grupp}
\alias{oppsum_grupp}
\title{Create a summary table for one variable grouped by another variable}
\usage{
oppsum_grupp(dataset, variable, by.var, export = NULL, norsk = NULL)
}
\arguments{
\item{dataset}{Your dataset (a data.frame-type object).}

\item{variable}{The variable for which you want summary statistics
(should ideally be a numeric variable).}

\item{by.var}{The variable that defines the groups over which the statistics
are calculated (should ideally be a categorical variable with not more than
10 unique categories).}

\item{export}{Should the result be made export-ready (TRUE/FALSE)?}

\item{norsk}{Results in Norwegian (NB)?}
}
\value{
A data.frame or, if export function is switched on (export=TRUE), a
printed out table formatted for easy export to MS Word.
}
\description{
This function allows you to create a table with summary statistics for one
variable over groups defined by another variable. The table will show the
following summary statistics: the number of nonmissing observations, the
average (mean), the median, the 25th and 75th percentiles, the standard
deviation, the minimum, the maximum, and the number of missing observations.
}
\details{
The result can be printed in an "export-ready" fashion, meaning
it can be copied and pasted into a Word document and there transformed into
a publication-quality table. The function will provide warnings in some
cases where user selections might be problematic.
}
\examples{
\dontrun{
# Load mtcars
data(mtcars)

# Simple
oppsum_grupp(dataset = mtcars,
variable = "drat",
by.var = "gear")

# With export function
result <- oppsum_grupp(dataset = mtcars,
variable = "mpg",
by.var = "cyl",
export = TRUE)
}

}
