% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oppsumtabell.R
\name{oppsumtabell}
\alias{oppsumtabell}
\title{Create univariate summary tables}
\usage{
oppsumtabell(dataset, variables, export = NULL)
}
\arguments{
\item{dataset}{Your dataset (a data.frame-type object).}

\item{variables}{The variables from your dataset for which you want
summary statistics. Specify them as a string or a string vector.}

\item{export}{Should the table be made export-ready (TRUE/FALSE)?}
}
\value{
A data.frame or, if export function is switched on (export=TRUE), a
printed out table formatted for easy export to MS Word.
}
\description{
This function allows you to create a simple univariate summary table for one
or more variables in a dataset. The table will show the following summary
statistics: the number of non-missing observations, the average (mean),
the median, the 25th and 75th percentiles, the standard deviation, the
minimum, the maximum, and the number of missing observations.
}
\details{
The result can be printed in an "export-ready" fashion, meaning
it can be copied and pasted into a Word document and there transformed into
a publication-quality table. The function will provide warnings in some
cases where user selections might be problematic.
}
\examples{
\dontrun{
# Loading mtcars dataset
data(mtcars)

# For a single variable
oppsumtabell(dataset = mtcars, variables = c("cyl"))

# Alternative
oppsumtabell(dataset = mtcars, variables = "cyl")

# For more variables
oppsumtabell(dataset = mtcars,
variables = c("cyl","mpg","disp"))

# Using export function
oppsumtabell(dataset = mtcars,
 variables = "mpg", export = TRUE)
}

}
