% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/practiceStatistics.R
\name{practiceStatistics}
\alias{practiceStatistics}
\title{An interactive dashboard featuring exercises & illustrations to practice statistical tests}
\usage{
practiceStatistics()
}
\description{
This function launches an interactive shiny-dashboard that allows users to practice calculating
and interpreting bivariate statistical tests (the t-test, the chi-squared test, and Pearson's
correlation coefficient). Users can let the dashboard generate random numbers to calculate test
statistics with. The dashboard also provides brief as well as detailed step-by-step solutions for
each type of test. In addition, the dashboard features a module to visualize statistical distributions
(the normal, t-, and chi-squared distributions) and critical values, a module to calculate p-values for
test statistics, a module to simulate the Central Limit Theorem, a module to illustrate the logic of
confidence intervals, and modules explaining fundamental materials such as mathematical concepts and
notation and the calculation of measures of central tendency (the mean and median) and dispersion
(the variance and standard deviation). The dashboard builds on (and can complement) Kellstedt & Whitten's
'Fundamentals of Political Science Research' (Cambridge Univ. Press, 2018).
}
\examples{
\dontrun{
practiceStatistics()
}

}
