% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visfactor.R
\name{visfactor}
\alias{visfactor}
\title{Show value-label correspondence for factor variables}
\usage{
visfactor(variable, dataset, detail = NULL, nrows = NULL)
}
\arguments{
\item{variable}{The variable from your dataset for which you want
summary statistics. Specify it as a string (e.g., "gndr").}

\item{dataset}{Your dataset (a data.frame-type object).}

\item{detail}{Do you want to get the 'detail' view (TRUE/FALSE)?}

\item{nrows}{How many rows do you want to get printed out if you
choose the detail-view (detail=TRUE). Must be a number.}
}
\value{
A data.frame.
}
\description{
This function shows you how the text labels and underlying numerical
values of a given factor-type variable correspond. You can use this
function in two ways: 1) in its default setting, the function returns
a simple table (more accurately, a data.frame) where each row is a
unique value of the factor variable. The table shows how numerical
values and text labels correspond. 2) You can choose the 'detail'
view, in which case the function prints out a table in which the
rows are actual observations from your dataset and shows you how
the factor variable's numerical values and text labels correspond
in your dataset. You can choose how many rows of the dataset are
printed, but the default is 15.
}
\examples{
\dontrun{
# Loading dataset
data(ess, package = "bst290")

# Default view
visfactor(dataset = ess, variable = "gndr")

# Detailed view
visfactor(dataset = ess, variable = "gndr", detail = TRUE)

# Detailed view, with 25 instead of 15 rows printed
visfactor(dataset = ess, variable = "gndr", detail = TRUE, nrows = 25)
}

}
